# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Minimal, production-ready Go API starter built with Fiber, GORM, JWT, and S3-compatible object storage.",
        "title": "Starter Go API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/ai-logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI Logs"
                ],
                "summary": "List my AI processing logs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.AIProcessingLogResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI Logs"
                ],
                "summary": "Record AI processing result",
                "parameters": [
                    {
                        "description": "AI log payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.CreateAILogRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.AIProcessingLogResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/v1/auth/change-password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Change my password",
                "parameters": [
                    {
                        "description": "Old + new password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.AuthResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new account",
                "parameters": [
                    {
                        "description": "Registration data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.AuthResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_ganiramadhan_starter-go_internal_dto.AIProcessingLogResponse": {
            "type": "object",
            "properties": {
                "attachment_id": {
                    "type": "string"
                },
                "confidence_score": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "extracted_amount": {
                    "type": "number"
                },
                "extracted_category": {
                    "type": "string"
                },
                "extracted_merchant": {
                    "type": "string"
                },
                "feature": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "latency_ms": {
                    "type": "integer"
                },
                "model_version": {
                    "type": "string"
                },
                "raw_response": {
                    "type": "object"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_ganiramadhan_starter-go_internal_dto.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.PaginationMeta"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "github_com_ganiramadhan_starter-go_internal_dto.AuthResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user": {
                    "$ref": "#/definitions/github_com_ganiramadhan_starter-go_internal_dto.UserResponse"
                }
            }
        },
...
        "github_com_ganiramadhan_starter-go_internal_dto.ChangePasswordRequest": {
            "type": "object",
            "required": [
                "current_password",
                "new_password"
            ],
            "properties": {
                "current_password": {
                    "type": "string",
                    "example": "oldSecret123"
                },
                "new_password": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 6,
                    "example": "newSecret123"
                }
            }
        },
...
        "github_com_ganiramadhan_starter-go_internal_dto.CreateAILogRequest": {
            "type": "object",
            "required": [
                "feature",
                "status"
            ],
            "properties": {
                "attachment_id": {
                    "type": "string"
                },
                "confidence_score": {
                    "type": "number",
                    "maximum": 1,
                    "minimum": 0
                },
                "error_message": {
                    "type": "string",
                    "maxLength": 2000
                },
                "extracted_amount": {
                    "type": "number",
                    "minimum": 0
                },
                "extracted_category": {
                    "type": "string",
                    "maxLength": 120
                },
                "extracted_merchant": {
                    "type": "string",
                    "maxLength": 255
                },
                "feature": {
                    "type": "string",
                    "enum": [
                        "categorize",
                        "scan_receipt",
                        "insights",
                        "suggest_budget",
                        "chat"
                    ]
                },
                "latency_ms": {
                    "type": "integer",
                    "minimum": 0
                },
                "model_version": {
                    "type": "string",
                    "maxLength": 64
                },
                "raw_response": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "success",
                        "failed"
                    ]
                }
            }
        },
...
        "github_com_ganiramadhan_starter-go_internal_dto.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                }
            }
        },
        "github_com_ganiramadhan_starter-go_internal_dto.PaginationMeta": {
            "type": "object",
            "properties": {
                "has_next": {
                    "type": "boolean",
                    "example": true
                },
                "has_previous": {
                    "type": "boolean",
                    "example": false
                },
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 100
                },
                "total_pages": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "github_com_ganiramadhan_starter-go_internal_dto.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "name": {
                    "type": "string",
                    "maxLength": 120,
                    "minLength": 2,
                    "example": "John Doe"
                },
                "password": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 6,
                    "example": "password123"
                }
            }
        },
...
        "github_com_ganiramadhan_starter-go_internal_dto.UserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "photo": {
                    "type": "string",
                    "example": "users/\u003cuuid\u003e/avatar.png"
                },
                "photo_url": {
                    "type": "string",
                    "example": "https://minio.local/starter/users/\u003cuuid\u003e/avatar.png?X-Amz-..."
                },
                "role": {
                    "type": "string",
                    "example": "user"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer {token}\" to authenticate.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}